refactor(core): split triple extraction service for issue #173
Extract rate limiter, error classification, result helpers, and LLM
provider calls into focused modules. Preserve TripleExtractionService
public API and behavior; replace Ollama model check any with typed guards.

Refs: https://github.com/jee1/memento/issues/173
Made-with: Cursor
